# Conflicts:
#	.github/workflows/release.yml
#	.github/workflows/test.yml
#	semver.txt
#	src/DotNetOutdated.Core/DotNetOutdated.Core.csproj
#	src/DotNetOutdated.Core/Services/ProjectAnalysisService.cs
#	src/DotNetOutdated/DotNetOutdated.csproj
#	test/DotNetOutdated.Tests/DotNetOutdated.Tests.csproj

+++ src/DotNetOutdated/Program.cs
@@ -28,15 +28,22 @@ namespace DotNetOutdated
        Name = "dotnet outdated",
        FullName = "A .NET Core global tool to list outdated Nuget packages.")]
    [VersionOptionFromMember(MemberName = nameof(GetVersion))]
-   internal class Program : CommandBase
+   internal class Program(
+        IFileSystem fileSystem,
+        IReporter reporter,
+        INuGetPackageResolutionService nugetService,
+        IProjectAnalysisService projectAnalysisService,
+        IProjectDiscoveryService projectDiscoveryService,
+        IDotNetPackageService dotNetPackageService,
+        ICentralPackageVersionManagementService centralPackageVersionManagementService) : CommandBase
    {
-      private readonly IFileSystem _fileSystem;
-      private readonly IReporter _reporter;
-      private readonly INuGetPackageResolutionService _nugetService;
-      private readonly IProjectAnalysisService _projectAnalysisService;
-      private readonly IProjectDiscoveryService _projectDiscoveryService;
-      private readonly IDotNetPackageService _dotNetPackageService;
-      private readonly ICentralPackageVersionManagementService _centralPackageVersionManagementService;
+      private readonly IFileSystem _fileSystem = fileSystem;
+      private readonly IReporter _reporter = reporter;
+      private readonly INuGetPackageResolutionService _nugetService = nugetService;
+      private readonly IProjectAnalysisService _projectAnalysisService = projectAnalysisService;
+      private readonly IProjectDiscoveryService _projectDiscoveryService = projectDiscoveryService;
+      private readonly IDotNetPackageService _dotNetPackageService = dotNetPackageService;
+      private readonly ICentralPackageVersionManagementService _centralPackageVersionManagementService = centralPackageVersionManagementService;
 
       [Option(CommandOptionType.NoValue, Description = "Specifies whether to include auto-referenced packages.",
           LongName = "include-auto-references")]
@@ -76,11 +83,11 @@ namespace DotNetOutdated
 
       [Option(CommandOptionType.MultipleValue, Description = "Specifies to only look at packages where the name contains the provided string. Culture and case insensitive. If provided multiple times, a single match is enough to include a package.",
           ShortName = "inc", LongName = "include")]
-      public List<string> FilterInclude { get; set; } = new List<string>();
+      public List<string> FilterInclude { get; set; } = [];
 
       [Option(CommandOptionType.MultipleValue, Description = "Specifies to only look at packages where the name does not contain the provided string. Culture and case insensitive. If provided multiple times, a single match is enough to exclude a package.",
           ShortName = "exc", LongName = "exclude")]
-      public List<string> FilterExclude { get; set; } = new List<string>();
+      public List<string> FilterExclude { get; set; } = [];
 
       [Option(CommandOptionType.SingleValue, Description = "Specifies the filename for a generated report. " +
                                                            "(Use the -of|--output-format option to specify the format. JSON by default.)",
@@ -122,14 +129,19 @@ namespace DotNetOutdated
       public LogLevel NuGetCredLogLevel { get; set; } = LogLevel.Warning;
       
       [Option(CommandOptionType.SingleValue, Description = "Specifies an optional runtime identifier to be used during the restore target when projects are analyzed. " +
-                                                           "More information available on https://learn.microsoft.com/en-us/dotnet/core/rid-catalog",
+                                                           "More information available on https://learn.microsoft.com/dotnet/core/rid-catalog",
          ShortName = "rt", LongName = "runtime")]
       public string Runtime { get; set; } = string.Empty;
+
+      [Option(CommandOptionType.SingleValue, Description = "The inclusive maximum package version to upgrade to." +
+                                                           "For example, a value of '8.0' would upgrade System.Text.Json 6.0.0 to the latest patch version of 8.0.x",
+         ShortName = "mv", LongName = "maximum-version")]
+      public string MaxVersion { get; set; } = string.Empty;
       
-        public static int Main(string[] args)
+      public static int Main(string[] args)
       {
          using var services = new ServiceCollection()
-                 .AddSingleton<IConsole>(PhysicalConsole.Singleton)
+                 .AddSingleton(PhysicalConsole.Singleton)
                  .AddSingleton<IReporter>(provider => new ConsoleReporter(provider.GetService<IConsole>()))
                  .AddSingleton<IFileSystem, FileSystem>()
                  .AddSingleton<IProjectDiscoveryService, ProjectDiscoveryService>()
@@ -156,18 +168,6 @@ namespace DotNetOutdated
           .GetCustomAttribute<AssemblyInformationalVersionAttribute>()
           .InformationalVersion;
 
-      public Program(IFileSystem fileSystem, IReporter reporter, INuGetPackageResolutionService nugetService, IProjectAnalysisService projectAnalysisService,
-          IProjectDiscoveryService projectDiscoveryService, IDotNetPackageService dotNetPackageService, ICentralPackageVersionManagementService centralPackageVersionManagementService)
-      {
-         _fileSystem = fileSystem;
-         _reporter = reporter;
-         _nugetService = nugetService;
-         _projectAnalysisService = projectAnalysisService;
-         _projectDiscoveryService = projectDiscoveryService;
-         _dotNetPackageService = dotNetPackageService;
-         _centralPackageVersionManagementService = centralPackageVersionManagementService;
-      }
-
       public async Task<int> OnExecute(CommandLineApplication app, IConsole console)
       {
          ArgumentNullException.ThrowIfNull(app);
@@ -191,13 +191,18 @@ namespace DotNetOutdated
             // Analyze the projects
             console.WriteLine("Analyzing project(s)...");
 
-            var projectLists = await Task.WhenAll(projectPaths.Select(path => _projectAnalysisService.AnalyzeProjectAsync(path, false, Transitive, TransitiveDepth, Runtime)));
+            var projectLists = new ConcurrentBag<List<Project>>();
+            await Parallel.ForEachAsync(projectPaths, async (path, _) =>
+            {
+                projectLists.Add(await _projectAnalysisService.AnalyzeProjectAsync(path, false, Transitive, TransitiveDepth, Runtime));
+            });
+
             var projects = projectLists.SelectMany(p => p).ToList();
 
             // Analyze the dependencies
             var outdatedProjects = await AnalyzeDependencies(projects, console).ConfigureAwait(false);
 
-            if (outdatedProjects.Any())
+            if (outdatedProjects.Count != 0)
             {
                // Report on the outdated dependencies
                ReportOutdatedDependencies(outdatedProjects, console);
@@ -280,9 +285,9 @@ namespace DotNetOutdated
                {
                   RunStatus status = null;
 
-                  if (!project.IsProjectSdkStyle())
+                  if (!project.IsProjectSdkStyle() && !package.IsVersionCentrallyManaged)
                   {
-                     console.WriteLine("Project format not SDK style, removing package before upgrade.");
+                     console.WriteLine("Project format not SDK style or centrally managed, removing package before upgrade.");
                      status = _dotNetPackageService.RemovePackage(project.ProjectFilePath, package.Name);
                   }
 
@@ -372,7 +377,7 @@ namespace DotNetOutdated
             {
                WriteTargetFramework(targetFramework, console);
 
-               var dependencies = targetFramework.Dependencies.ToList();
+               var dependencies = targetFramework.Dependencies;
 
                int[] columnWidths = dependencies.DetermineColumnWidths();
 
@@ -414,15 +419,7 @@ namespace DotNetOutdated
 
          console.WriteLine("Analyzing dependencies...");
 
-         var tasks = new Task[projects.Count];
-
-         for (var index = 0; index < projects.Count; index++)
-         {
-            var project = projects[index];
-            tasks[index] = AddOutdatedProjectsIfNeeded(project, outdatedProjects);
-         }
-
-         await Task.WhenAll(tasks).ConfigureAwait(false);
+         await Parallel.ForEachAsync(projects, async (project, _) => await AddOutdatedProjectsIfNeeded(project, outdatedProjects));
 
          return outdatedProjects
              .OrderBy(p => p.Name)
@@ -462,25 +459,51 @@ namespace DotNetOutdated
       {
          var outdatedDependencies = new ConcurrentBag<AnalyzedDependency>();
 
-         var deps = targetFramework.Dependencies.Where(d => this.IncludeAutoReferences || !d.IsAutoReferenced);
+         IEnumerable<Dependency> deps = targetFramework.Dependencies.Values;
+         
+         if (!this.IncludeAutoReferences)
+         {
+            deps = deps.Where(d => !d.IsAutoReferenced);
+         }
 
-         if (FilterInclude.Any())
+         if (FilterInclude.Count != 0)
             deps = deps.Where(AnyIncludeFilterMatches);
 
-         if (FilterExclude.Any())
+         if (FilterExclude.Count != 0)
             deps = deps.Where(NoExcludeFilterMatches);
+
+         NuGetVersion maximumVersion = null;
+
+         if (!string.IsNullOrEmpty(MaxVersion))
+         {
+             if (!Version.TryParse(MaxVersion, out var maxVersion))
+             {
+                 throw new CommandValidationException($"The specified maximum version '{MaxVersion}' is not a valid version string.");
+             }
+
+             // NuGetVersion normalizes no build or revision to 0, when we actually want
+             // those to mean "any version", so we need to force them to int.MaxValue.
+             if (maxVersion.Build is -1 || maxVersion.Revision is -1)
+             {
+                 var build = maxVersion.Build == -1 ? int.MaxValue : maxVersion.Build;
+                 var revision = maxVersion.Revision == -1 ? int.MaxValue : maxVersion.Revision;
+                 maxVersion = new(maxVersion.Major, maxVersion.Minor, build, revision);
+             }
+
+             maximumVersion = new NuGetVersion(maxVersion);
+         }
 
          var dependencies = deps.OrderBy(dependency => dependency.IsTransitive)
              .ThenBy(dependency => dependency.Name)
-             .ToList();
+             .ToArray();
 
-         var tasks = new Task[dependencies.Count];
+         var tasks = new Task[dependencies.Length];
 
-         for (var index = 0; index < dependencies.Count; index++)
+         for (var index = 0; index < dependencies.Length; index++)
          {
             var dependency = dependencies[index];
 
-            tasks[index] = this.AddOutdatedDependencyIfNeeded(project, targetFramework, dependency, outdatedDependencies);
+            tasks[index] = AddOutdatedDependencyIfNeeded(project, targetFramework, dependency, maximumVersion, outdatedDependencies);
          }
 
          await Task.WhenAll(tasks).ConfigureAwait(false);
@@ -489,10 +512,29 @@ namespace DotNetOutdated
             outdatedFrameworks.Add(new AnalyzedTargetFramework(targetFramework.Name, outdatedDependencies));
       }
 
-      private async Task AddOutdatedDependencyIfNeeded(Project project, TargetFramework targetFramework, Dependency dependency, ConcurrentBag<AnalyzedDependency> outdatedDependencies)
+      private async Task AddOutdatedDependencyIfNeeded(
+          Project project,
+          TargetFramework targetFramework,
+          Dependency dependency,
+          NuGetVersion maximumVersion,
+          ConcurrentBag<AnalyzedDependency> outdatedDependencies)
       {
          var referencedVersion = dependency.ResolvedVersion;
+         var versionRange = dependency.VersionRange;
          NuGetVersion latestVersion = null;
+
+        if (maximumVersion is not null &&
+            (versionRange.MaxVersion is null || maximumVersion > versionRange.MaxVersion))
+        {
+            // Patch the version range to include the user-specified maximum
+            versionRange = new(
+                versionRange.MinVersion,
+                versionRange.IsMinInclusive,
+                maximumVersion,
+                includeMaxVersion: true,
+                versionRange.Float,
+                versionRange.OriginalString);
+        }
 
          if (referencedVersion != null)
          {
@@ -500,7 +542,7 @@ namespace DotNetOutdated
                 dependency.Name,
                 referencedVersion,
                 project.Sources,
-                dependency.VersionRange,
+                versionRange,
                 VersionLock,
                 Prerelease,
                 PrereleaseLabel,
@@ -520,7 +562,7 @@ namespace DotNetOutdated
                    dependency.Name,
                    referencedVersion,
                    project.Sources,
-                   dependency.VersionRange,
+                   versionRange,
                    VersionLock,
                    Prerelease,
                    PrereleaseLabel,
